feat: add track route
