+++ src/add-router-tracking.ts
@@ -1,28 +1,5 @@
-import type { RouteLocationNormalizedGeneric } from "vue-router";
-import query from "./gtag/query";
-import { getSettings } from "./settings";
-
-function isRouteExcluded(route: RouteLocationNormalizedGeneric) {
-  const { excludedRoutes } = getSettings();
-
-  return excludedRoutes?.some(({ name, path } = {}) => {
-    return (name && name === route.name) || (path && path === route.path);
-  });
-}
-
-function trackRoute(route: RouteLocationNormalizedGeneric) {
-  const { onBeforeTrack, onAfterTrack } = getSettings();
-
-  if (isRouteExcluded(route)) {
-    return;
-  }
-
-  onBeforeTrack?.();
-
-  query("event", "page_view", { page_path: route.path });
-
-  onAfterTrack?.();
-}
+import { getSettings } from "@/settings";
+import trackRoute from "@/track-route";
 
 export default async function addRouterTracking(): Promise<void> {
   const { router } = getSettings();
